Keep only the main definitions in Quote, move specific functions to specific modules, break cycles
