+++ language-plutus-core/src/Language/PlutusCore/Error.hs
@@ -1,17 +1,42 @@
 {-# LANGUAGE ConstrainedClassMethods #-}
 {-# LANGUAGE DeriveAnyClass          #-}
 {-# LANGUAGE FlexibleInstances       #-}
+{-# LANGUAGE OverloadedStrings       #-}
 
 module Language.PlutusCore.Error ( Error (..)
+                                 , RenameError (..)
+                                 , TypeError (..)
                                  , IsError (..)
                                  ) where
 
 import           Language.PlutusCore.Lexer
+import           Language.PlutusCore.Name
 import           Language.PlutusCore.Normalize
 import           Language.PlutusCore.PrettyCfg
-import           Language.PlutusCore.Renamer
-import           Language.PlutusCore.TypeSynthesis
+import           Language.PlutusCore.Type
 import           PlutusPrelude
+
+-- | A 'RenameError' is thrown when a free variable is encountered during
+-- rewriting.
+data RenameError a = UnboundVar (Name a)
+                   | UnboundTyVar (TyName a)
+                   deriving (Generic, NFData)
+
+instance (PrettyCfg a) => PrettyCfg (RenameError a) where
+    prettyCfg cfg (UnboundVar n@(Name loc _ _)) = "Error at" <+> prettyCfg cfg loc <> ". Variable" <+> prettyCfg cfg n <+> "is not in scope."
+    prettyCfg cfg (UnboundTyVar n@(TyName (Name loc _ _))) = "Error at" <+> prettyCfg cfg loc <> ". Type variable" <+> prettyCfg cfg n <+> "is not in scope."
+
+data TypeError a = InternalError -- ^ This is thrown if builtin lookup fails
+                 | KindMismatch a (Type TyNameWithKind ()) (Kind ()) (Kind ())
+                 | TypeMismatch a (Term TyNameWithKind NameWithType ()) (Type TyNameWithKind ()) (Type TyNameWithKind ())
+                 | OutOfGas
+                 deriving (Generic, NFData)
+
+instance (PrettyCfg a) => PrettyCfg (TypeError a) where
+    prettyCfg _ InternalError               = "Internal error."
+    prettyCfg cfg (KindMismatch x ty k k')  = "Kind mismatch at" <+> prettyCfg cfg x <+> "in type" <+> squotes (prettyCfg cfg ty) <> ". Expected kind" <+> squotes (pretty k) <+> ", found kind" <+> squotes (pretty k')
+    prettyCfg cfg (TypeMismatch x t ty ty') = "Type mismatch at" <+> prettyCfg cfg x <+> "in term" <> hardline <> indent 2 (squotes (prettyCfg cfg t)) <> "." <> hardline <> "Expected type" <> hardline <> indent 2 (squotes (prettyCfg cfg ty)) <> "," <> hardline <> "found type" <> hardline <> indent 2 (squotes (prettyCfg cfg ty'))
+    prettyCfg _ OutOfGas                    = "Type checker ran out of gas."
 
 data Error a = ParseError (ParseError a)
              | RenameError (RenameError a)
