+++ language-plutus-core/src/Language/PlutusCore/Renamer.hs
@@ -1,26 +1,23 @@
-{-# LANGUAGE DeriveAnyClass     #-}
 {-# LANGUAGE DerivingStrategies #-}
 {-# LANGUAGE FlexibleContexts   #-}
 {-# LANGUAGE FlexibleInstances  #-}
-{-# LANGUAGE OverloadedStrings  #-}
 
 module Language.PlutusCore.Renamer ( rename
                                    , annotate
                                    , annotateST
                                    , annotateTermST
-                                   , NameWithType (..)
-                                   , RenamedType
-                                   , RenamedTerm
-                                   , TyNameWithKind (..)
-                                   , RenameError (..)
+                                   , annotateProgramQ
+                                   , annotateTermQ
                                    , TypeState (..)
+                                   , RenameError (..)
                                    ) where
 
 import           Control.Monad.Except
 import           Control.Monad.State.Lazy
-import qualified Data.IntMap                   as IM
+import qualified Data.IntMap               as IM
+import           Language.PlutusCore.Error
 import           Language.PlutusCore.Name
-import           Language.PlutusCore.PrettyCfg
+import           Language.PlutusCore.Quote
 import           Language.PlutusCore.Type
 import           Lens.Micro
 import           PlutusPrelude
@@ -42,32 +39,22 @@ instance Monoid (TypeState a) where
 
 type TypeM a = StateT (TypeState a) (Either (RenameError a))
 
-type RenamedTerm a = Term TyNameWithKind NameWithType a
-newtype NameWithType a = NameWithType (Name (a, RenamedType a))
-    deriving (Functor, Generic)
-    deriving newtype NFData
-type RenamedType a = Type TyNameWithKind a
-newtype TyNameWithKind a = TyNameWithKind { unTyNameWithKind :: TyName (a, Kind a) }
-    deriving (Eq, Functor, Generic)
-    deriving newtype NFData
+-- | Annotate a PLC program, so that all names are annotated with their types/kinds.
+annotateProgramQ :: (MonadError (Error a) m, MonadQuote m) => Program TyName Name a -> m (Program TyNameWithKind NameWithType a)
+annotateProgramQ (Program a v t) = Program a v <$> annotateTermQ t
 
-instance PrettyCfg (TyNameWithKind a) where
-    prettyCfg cfg@(Configuration _ True) (TyNameWithKind (TyName tn@(Name (_, k) _ _))) = parens (prettyCfg cfg tn <+> ":" <+> pretty k)
-    prettyCfg cfg@(Configuration _ False) (TyNameWithKind tn) = prettyCfg cfg tn
-
-instance PrettyCfg (NameWithType a) where
-    prettyCfg cfg@(Configuration _ True) (NameWithType n@(Name (_, ty) _ _)) = parens (prettyCfg cfg n <+> ":" <+> prettyCfg cfg ty)
-    prettyCfg cfg@(Configuration _ False) (NameWithType n) = prettyCfg cfg n
-
--- | A 'RenameError' is thrown when a free variable is encountered during
--- rewriting.
-data RenameError a = UnboundVar (Name a)
-                   | UnboundTyVar (TyName a)
-                   deriving (Generic, NFData)
-
-instance (PrettyCfg a) => PrettyCfg (RenameError a) where
-    prettyCfg cfg (UnboundVar n@(Name loc _ _)) = "Error at" <+> prettyCfg cfg loc <> ". Variable" <+> prettyCfg cfg n <+> "is not in scope."
-    prettyCfg cfg (UnboundTyVar n@(TyName (Name loc _ _))) = "Error at" <+> prettyCfg cfg loc <> ". Type variable" <+> prettyCfg cfg n <+> "is not in scope."
+-- | Annotate a PLC term, so that all names are annotated with their types/kinds.
+annotateTermQ :: forall m a . (MonadError (Error a) m, MonadQuote m) => Term TyName Name a -> m (Term TyNameWithKind NameWithType a)
+annotateTermQ t = convertErrors asError $ do
+    (ts, t') <- liftEither $ annotateTermST t
+    updateMaxU ts
+    pure t'
+        where
+            updateMaxU (TypeState _ tys) = do
+                nextU <- liftQuote get
+                let tsMaxU = maybe 0 (fst . fst) (IM.maxViewWithKey tys)
+                let maxU = unUnique nextU - 1
+                liftQuote $ put $ Unique $ max maxU tsMaxU
 
 -- | Annotate a program with type/kind information at all bound variables,
 -- failing if we encounter a free variable.
