+++ language-plutus-core/src/Language/PlutusCore/Type.hs
@@ -1,7 +1,8 @@
-{-# LANGUAGE DeriveAnyClass    #-}
-{-# LANGUAGE FlexibleContexts  #-}
-{-# LANGUAGE OverloadedStrings #-}
-{-# LANGUAGE TypeFamilies      #-}
+{-# LANGUAGE DeriveAnyClass     #-}
+{-# LANGUAGE DerivingStrategies #-}
+{-# LANGUAGE FlexibleContexts   #-}
+{-# LANGUAGE OverloadedStrings  #-}
+{-# LANGUAGE TypeFamilies       #-}
 
 module Language.PlutusCore.Type ( Term (..)
                                 , Value
@@ -19,6 +20,11 @@ module Language.PlutusCore.Type ( Term (..)
                                 -- * Helper functions
                                 , tyLoc
                                 , termLoc
+                                -- * Renamed types and terms
+                                , NameWithType (..)
+                                , RenamedType
+                                , RenamedTerm
+                                , TyNameWithKind (..)
                                 ) where
 
 import qualified Data.ByteString.Lazy           as BSL
@@ -26,6 +32,7 @@ import           Data.Functor.Foldable
 import           Instances.TH.Lift              ()
 import           Language.Haskell.TH.Syntax     (Lift)
 import           Language.PlutusCore.Lexer.Type
+import           Language.PlutusCore.Name
 import           Language.PlutusCore.PrettyCfg
 import           PlutusPrelude
 
@@ -234,3 +241,21 @@ instance (PrettyCfg (f a)) => PrettyCfg (Type f a) where
         a (TyBuiltinF _ n)    = parens ("con" <+> pretty n)
         a (TyIntF _ n)        = parens ("con" <+> pretty n)
         a (TyLamF _ n k t)    = parens ("lam" <+> prettyCfg cfg n <+> pretty k <+> t)
+
+type RenamedTerm a = Term TyNameWithKind NameWithType a
+newtype NameWithType a = NameWithType (Name (a, RenamedType a))
+    deriving (Functor, Generic)
+    deriving newtype NFData
+
+type RenamedType a = Type TyNameWithKind a
+newtype TyNameWithKind a = TyNameWithKind { unTyNameWithKind :: TyName (a, Kind a) }
+    deriving (Eq, Functor, Generic)
+    deriving newtype NFData
+
+instance PrettyCfg (TyNameWithKind a) where
+    prettyCfg cfg@(Configuration _ True) (TyNameWithKind (TyName tn@(Name (_, k) _ _))) = parens (prettyCfg cfg tn <+> ":" <+> pretty k)
+    prettyCfg cfg@(Configuration _ False) (TyNameWithKind tn) = prettyCfg cfg tn
+
+instance PrettyCfg (NameWithType a) where
+    prettyCfg cfg@(Configuration _ True) (NameWithType n@(Name (_, ty) _ _)) = parens (prettyCfg cfg n <+> ":" <+> prettyCfg cfg ty)
+    prettyCfg cfg@(Configuration _ False) (NameWithType n) = prettyCfg cfg n
